refactor(src/pages): switched to page router
removed src/app dir. added pages dir for all pages. moved AppRouter.jsx to src dir. refactored imports and import alias.

reason: Ekum feels more comfortable working with page router.
